Bug fixes, changes and additions.
The instance methods read and write were documented to return the total number of bytes read/written, but were actually returning the number of objects.
Removed assert checks on the file handle, since the PHYSFS functions will check that anyway.
Added new versions of read and write that take void* pointers rather than ubyte buffers. This allows using the methods in the same manner as the PHYSFS functions to read single byte values or other integrals without worrying about manipulating bytes in a buffer.
Added readLE/BE and writeLE/BE templated methods to wrap the little-endian and big-endian read/write functions for 16-/32-/64-bit types.